Make segment form draggable and make only segment type tab swipable
