+++ features/routines/src/main/java/com/enricog/features/routines/detail/segment/ui_components/SegmentFormScene.kt
@@ -2,9 +2,9 @@ package com.enricog.features.routines.detail.segment.ui_components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.Orientation
-import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.gestures.draggable
+import androidx.compose.foundation.gestures.rememberDraggableState
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -13,9 +13,7 @@ import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.FractionalThreshold
 import androidx.compose.material.rememberSwipeableState
-import androidx.compose.material.swipeable
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.ExperimentalComposeUiApi
@@ -27,13 +25,11 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import com.enricog.core.compose.api.extensions.stringResourceOrNull
-import com.enricog.core.compose.api.modifiers.horizontalListItemSpacing
 import com.enricog.data.routines.api.entities.Segment
 import com.enricog.data.routines.api.entities.TimeType
 import com.enricog.entities.Seconds
 import com.enricog.features.routines.R
 import com.enricog.features.routines.detail.segment.models.SegmentField
-import com.enricog.features.routines.detail.ui_components.TimeTypeChip
 import com.enricog.ui.components.button.TempoButton
 import com.enricog.ui.components.button.TempoButtonColor
 import com.enricog.ui.components.textField.TempoTextField
@@ -52,28 +48,27 @@ internal fun SegmentFormScene(
     onSegmentTimeChange: (Seconds) -> Unit,
     onSegmentTimeTypeChange: (TimeType) -> Unit,
     onSegmentConfirmed: () -> Unit
-) = BoxWithConstraints {
+) {
     val keyboardController = LocalSoftwareKeyboardController.current
     val (segmentNameRef, segmentTimeRef) = remember { FocusRequester.createRefs() }
 
-    val width = constraints.maxWidth.toFloat()
     val swipeState = rememberSwipeableState(initialValue = segment.type) {
         onSegmentTimeTypeChange(it)
         true
     }
-    val anchors =
-        timeTypes.mapIndexed { index, timeType -> width / (index + 1) to timeType }.toMap()
+    val draggableState = rememberDraggableState {
+        swipeState.performDrag(it)
+    }
 
     Column(
         modifier = Modifier
             .testTag(SegmentFormSceneTestTag)
             .fillMaxSize()
             .background(TempoTheme.colors.background)
-            .swipeable(
-                state = swipeState,
-                anchors = anchors,
-                thresholds = { _, _ -> FractionalThreshold(0.5f) },
-                orientation = Orientation.Horizontal
+            .draggable(
+                state = draggableState,
+                orientation = Orientation.Horizontal,
+                onDragStopped = { swipeState.performFling(it) }
             )
     ) {
         Column(
@@ -82,7 +77,10 @@ internal fun SegmentFormScene(
                 .weight(1f)
                 .verticalScroll(rememberScrollState(0))
         ) {
-            SelectableTimeType(
+
+            SegmentTypeTab(
+                modifier = Modifier.fillMaxWidth(),
+                state = swipeState,
                 timeTypes = timeTypes,
                 selected = segment.type,
                 onSelectChange = onSegmentTimeTypeChange
@@ -133,30 +131,4 @@ internal fun SegmentFormScene(
     }
 }
 
-@Composable
-private fun SelectableTimeType(
-    timeTypes: List<TimeType>,
-    selected: TimeType,
-    onSelectChange: (TimeType) -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .padding(TempoTheme.dimensions.spaceS)
-            .fillMaxWidth()
-    ) {
-        timeTypes.mapIndexed { index, timeType ->
-            TimeTypeChip(
-                value = timeType,
-                isSelected = selected == timeType,
-                onClick = onSelectChange,
-                modifier = Modifier
-                    .horizontalListItemSpacing(
-                        itemPosition = index,
-                        spacing = TempoTheme.dimensions.spaceS,
-                        includeEdge = true
-                    )
-                    .weight(weight = 1f, fill = true)
-            )
-        }
-    }
-}
+
